bug 7811: Start domR/Dhcp if it's not running when do userVmReboot status 7811: resolved fixed
Other fixes:
* vmExpunge: cleanup LB/PF rules after vm was marked as Expunging in the DB to avoid the situation when user recovers a vm in the middle of expunge job.

+++ server/src/com/cloud/network/lb/LoadBalancingRulesManagerImpl.java
@@ -208,10 +208,15 @@ public class LoadBalancingRulesManagerImpl implements LoadBalancingRulesManager,
                 _lb2VmMapDao.persist(map);
                 s_logger.debug("Set load balancer rule for revoke: rule id " + loadBalancerId + ", vmId " + instanceId);
             }
-
-            applyLoadBalancerConfig(loadBalancerId);
-            _lb2VmMapDao.remove(loadBalancerId, instanceIds, null);
-            s_logger.debug("Load balancer rule id " + loadBalancerId + " is removed for vms " + instanceIds);
+            
+            if (applyLoadBalancerConfig(loadBalancerId)) {
+                _lb2VmMapDao.remove(loadBalancerId, instanceIds, null);
+                s_logger.debug("Load balancer rule id " + loadBalancerId + " is removed for vms " + instanceIds);
+            } else {
+                s_logger.warn("Failed to remove load balancer rule id " + loadBalancerId + " for vms " + instanceIds);
+                throw new CloudRuntimeException("Failed to remove load balancer rule id " + loadBalancerId + " for vms " + instanceIds);
+            }
+            
         } catch (ResourceUnavailableException e) {
             s_logger.warn("Unable to apply the load balancer config because resource is unavaliable.", e);
             return false;
@@ -418,7 +423,7 @@ public class LoadBalancingRulesManagerImpl implements LoadBalancingRulesManager,
         String dstIp = null;
         for (LoadBalancerVMMapVO lbVmMap : lbVmMaps) {
             UserVm vm = _vmDao.findById(lbVmMap.getInstanceId());
-            Nic nic = _nicDao.findByInstanceIdAndNetworkId(lb.getNetworkId(), vm.getId());
+            Nic nic = _nicDao.findByInstanceIdAndNetworkIdIncludingRemoved(lb.getNetworkId(), vm.getId());
             dstIp = nic.getIp4Address();
             LbDestination lbDst = new LbDestination(lb.getDefaultPortStart(), lb.getDefaultPortEnd(), dstIp, lbVmMap.isRevoke());
             dstList.add(lbDst);
